Rename `cupyx.runtime` submodule
